Renumber the stack migrations to 354-356 because upstream main took 353 for
the audit event table, and carry the re-request review type into the
codeowner review checks.

+++ modules/setting/repository.go
@@ -113,7 +113,7 @@ var (
 			AllowedTypes     string
 			DefaultPagingNum int
 			FileMaxSize      int64
-			MaxFiles         int64
+			MaxFiles         int
 		} `ini:"repository.release"`
 
 		Signing struct {
@@ -228,7 +228,7 @@ var (
 			AllowedTypes     string
 			DefaultPagingNum int
 			FileMaxSize      int64
-			MaxFiles         int64
+			MaxFiles         int
 		}{
 			AllowedTypes:     "",
 			DefaultPagingNum: 10,
